soc: qcom: bam_dmux: Sync SSR with disconnect/reconnect
The disconnect and reconnect bam events access the hardware which may not
be permitted durring subsystem restart due to the risk of a stalled access.

Put an explicit sync event between SSR and the disconnect/reconnect events
to eliminate any risk of hardware accesses which are not permitted.

CRs-Fixed: 771505
Change-Id: I206c1cf6f18955320aa20c1dac56eb2173675c4c
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>

Conflicts:
	arch/arm/mach-msm/bam_dmux.c